Clean up package name and source files
Use the buildpacks namespace and gather together the transform and
replace operations defined on a SourceFile.
