+++ cmd/root.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 
-	scafall "github.com/AidanDelaney/scafall/pkg"
+	scafall "github.com/buildpacks/scafall/pkg"
 )
 
 const (
