+++ pkg/internal/source_file.go
@@ -1,11 +1,46 @@
 package internal
 
 import (
+	"fmt"
+	"io/fs"
+	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 
 	t "github.com/coveooss/gotemplate/v3/template"
 )
+
+type SourceFile struct {
+	FilePath    string
+	FileContent string
+	FileMode    fs.FileMode
+}
+
+func (s SourceFile) Transform(inputDir string, outputDir string, vars map[string]string) error {
+	outputFile, err := s.Replace(vars)
+	if err != nil {
+		return err
+	}
+
+	dstDir := filepath.Join(outputDir, filepath.Dir(outputFile.FilePath))
+	mkdirErr := os.MkdirAll(dstDir, 0744)
+	if mkdirErr != nil {
+		return fmt.Errorf("failed to create target directory %s", dstDir)
+	}
+
+	outputPath := filepath.Join(outputDir, outputFile.FilePath)
+	if outputFile.FileContent == "" {
+		inputPath := filepath.Join(inputDir, s.FilePath)
+		mvErr := os.Rename(inputPath, outputPath)
+		if mvErr != nil {
+			return fmt.Errorf("failed to rename %s to %s", s.FilePath, outputFile.FilePath)
+		}
+	} else {
+		os.WriteFile(outputPath, []byte(outputFile.FileContent), outputFile.FileMode|0600)
+	}
+	return nil
+}
 
 func replaceUnknownVars(vars map[string]string, content string) string {
 	regex := regexp.MustCompile(`{{[ \t]*\.\w+`)
@@ -21,7 +56,7 @@ func replaceUnknownVars(vars map[string]string, content string) string {
 	return transformed
 }
 
-func Replace(vars map[string]string, file SourceFile) (SourceFile, error) {
+func (s SourceFile) Replace(vars map[string]string) (SourceFile, error) {
 	opts := t.DefaultOptions().
 		Set(t.Overwrite, t.Sprig, t.StrictErrorCheck, t.AcceptNoValue).
 		Unset(t.Razor)
@@ -34,7 +69,7 @@ func Replace(vars map[string]string, file SourceFile) (SourceFile, error) {
 		return SourceFile{}, err
 	}
 
-	filePath := replaceUnknownVars(vars, file.FilePath)
+	filePath := replaceUnknownVars(vars, s.FilePath)
 	transformedFilePath, err := template.ProcessContent(filePath, "")
 	if err != nil {
 		return SourceFile{}, err
@@ -42,8 +77,8 @@ func Replace(vars map[string]string, file SourceFile) (SourceFile, error) {
 	transformedFilePath = strings.ReplaceAll(transformedFilePath, ReplacementDelimiter, "{{")
 
 	transformedFileContent := ""
-	if file.FileContent != "" {
-		fileContent := replaceUnknownVars(vars, file.FileContent)
+	if s.FileContent != "" {
+		fileContent := replaceUnknownVars(vars, s.FileContent)
 		transformedFileContent, err = template.ProcessContent(fileContent, "")
 		if err != nil {
 			return SourceFile{}, err
